Setup and Teardown - Setup is done in the constructor. - Cleanup is done using Dispose with IDisposable. What if it's Async? - Using the IAsyncLifetime interface from XUnit.
ITestOutputHelper is an interface that will be injected by XUnit so you can print information that appears in the test console.

Parametrization
To test multiple values with the same test.
1. You first use the Theory Attribute.
2. Add the input and expected as parameters.
3. Then add data with the InlineData Attribute for each case.
```
    [Theory]
    [InlineData(8,4,12)]

    [InlineData(0,0,0)]
    [InlineData(-10,-11,-21)]
    [InlineData(10,-11,-1)]
    public void Add_ShouldAdd_WhenTwoInts(int a, int b, int expected)
    {
        //Arrange
        _outputHelper.WriteLine("This is the Add Test");

        //Act
        var result = _sut.Add(a, b);

        //Assert
        Assert.Equal(expected, result);
    }
```

Ignoring Tests
Add this parameter to the Theory or Fact attribute: [Theory(Skip = "Skiping this for test.")]
You can also do it in the InlineData attribute.