Created an API aplication for testing Endpoints. Used Fakes tu test a service.
Fakes
Fakes is the idea of implementing a the Interface that is needed to be injected to do the tests. Its falls apart when you have multiple tests with different requirementes. The Fake can get bigger and have problems of it's own. It's good. But might be too slow or not enough for harder cases.