Implemented the same Service Test but using Mock instead of the fake class.
Mock
It's a simpler representation of the dependency that you configure to have an especific behavior.
```
//This creates the repository.
_repository = Substitute.For<ICustomerRepository>();
//This tells it what to do:
_repository.GetAllAsync().Returns(Array.Empty<Customer>());
```
NSubstite creates an in memory implementation of what to do, and it's actually called via a Proxy object.
NSubstitute because Moq requieres too much ceremony, and most experts in C# agree. Also Moq has it's call home issue.