Added Testing for the Service Logging.
Testing ILogger
ILogger is not easy to test because is built as extension methods that NSubstitute won't be able to replicate, nor Moq nor any mockinf. At the end you shold create an adapter so you can check for the calls and parameters sent.
You create an ILoggerAdapter<T> interface with the methods you need to test, they have to have the same signature as the real methods, then implement the LoggerAdapter with a private ILogger that just passes the calls to the real one. This way you can "intercept" the calls.
Then you register it as a TransientService in the DI container.