You might want to consider creating one test class per HTTP command. We created one class per each.
If you do this, you will be creating one web server per class. Wich is not bad, but if you want to use the same one you will use Xinit ICollectionFixture so only one is used for all tests. First create a class that implements the ICollectionFicture Interface, with the WebApplicationFactoryClass.
```
[CollectionDefinition("HttpServerCollection")]
public class TestCollection : ICollectionFixture<WebApplicationFactory<IApiMarker>>{ }
```
Then decorate your classes with the Collection Attribute, remember to use the exact same name, and remove the IClassFixture interface.
```
[Collection("HttpServerCollection")]
public class GetCustomerControllerTests
```
With this, the same WebApplicationFactory will be shared with all the tests.