Faking API with WireMock When you have external APIs that you don't control, you might not want to call the real API every test. Maybe because it has a call limit, maybe because you want different responses. We can do this with Wire Mock. Is similar to mocking but with real http calls. To use it you add the library to the project. ``` dotnet add package WireMock.Net ``` TIP: When I tried to install there was an error saying Microsoft.CodeAnalysis was not the right version. It asked for 4.8.0-3.final and I had 4.8.0. It looks that there is a prerelease, after looking for it I found: ``` dotnet add package Microsoft.CodeAnalysis --version 4.8.0-3.final ``` This doesn't solved the problem. The solution was to install WireMock 1.5.47.
Then you simply define what calls should WireMock Answer, with what response.
```
//Program.cs
var wiremockServer = WireMockServer.Start();
Console.WriteLine($"Wiremock is now running on: {wiremockServer.Url}");

wiremockServer.Given(Request.Create().WithPath("/example").UsingGet())
    .ResponsWidth(Response.Create().WithBody("Whis is coming from WireMock").WithSatatusCode(200));

Console.ReadKey();
wiremockServer.Dispose();
```
In the example we create a class to simulate a user check in Github. It initiates with the CustomerApiFactory and stops when it is turned off.
Then you add the URL in the Services definition with the URL pointing to a exposed variable in the class with it. And with that you can start testing against this API.